requirements.txt
asgiref==3.9.1
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.5
django-cors-headers==4.7.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
google-auth==2.40.3
idna==3.10
pillow==11.3.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyJWT==2.10.1
requests==2.32.4
rsa==4.9.1
sqlparse==0.5.3
urllib3==2.5.0
